+++ crates/cli/src/wasm_bindgen_test_runner/server.rs
@@ -260,15 +260,32 @@ pub(crate) fn spawn(
             // Wasm module.
             document.getElementById('output').textContent = "Loading Wasm module...";
 
+            const nocapture = {nocapture};
+            const wrap = method => {{
+                const og = console[method];
+                console[method] = function (...args) {{
+                    if (nocapture) {{
+                        window.__wbg_test_output_writeln(args);
+                    }}
+                    og(args[0]);
+                }};
+            }};
+
+            wrap("debug");
+            wrap("log");
+            wrap("info");
+            wrap("warn");
+            wrap("error");
+
+            window.__wbg_test_output = "";
+            window.__wbg_test_output_writeln = function (line) {{
+                window.__wbg_test_output += line + "\n";
+                document.getElementById("output").textContent = window.__wbg_test_output;
+            }}
+
             async function main(test) {{
                 const wasm = await init('./{module}_bg.wasm');
-
                 const cx = new Context();
-                window.on_console_debug = __wbgtest_console_debug;
-                window.on_console_log = __wbgtest_console_log;
-                window.on_console_info = __wbgtest_console_info;
-                window.on_console_warn = __wbgtest_console_warn;
-                window.on_console_error = __wbgtest_console_error;
 
                 {args}